src/main/java/com/cognizant/User/model/User.java
package com.cognizant.User.model;

import javax.persistence.*;
import java.util.Objects;

@Entity
@Table(name="users")
public class User {

    @Id
    private int id;

    @Column(name="username")
    private String username;

    @Column(name="profile_img")
    private String profileImg;

    public User() {
    }

    public User(int id, String username, String profileImg) {
        this.id = id;
        this.username = username;
        this.profileImg = profileImg;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getProfileImg() {
        return profileImg;
    }

    public void setProfileImg(String profileImg) {
        this.profileImg = profileImg;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        User user = (User) o;
        return id == user.id && username.equals(user.username) && Objects.equals(profileImg, user.profileImg);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, username, profileImg);
    }

    @Override
    public String toString() {
        return "User{" +
                "id=" + id +
                ", username='" + username + '\'' +
                ", profileImg='" + profileImg + '\'' +
                '}';
    }
}
